fix slice
