+++ rfc9535/src/suite.rs
@@ -1,6 +1,8 @@
 use colored::Colorize;
-use jsonpath_rust::JsonPathParserError;
+use jsonpath_rust::{JsonPath, JsonPathParserError};
 use serde_json::Value;
+use std::str::FromStr;
+use crate::console::TestResult;
 
 type SkippedCases = usize;
 
@@ -32,6 +34,43 @@ pub fn get_suite() -> Result<(Vec<TestCase>, SkippedCases), std::io::Error> {
         skipped_cases,
     ))
 }
+pub fn handle_test_case(case: &TestCase) -> TestResult {
+    let js_path: Result<JsonPath<Value>, _> = JsonPath::from_str(case.selector.as_str());
+
+    if case.invalid_selector {
+        if js_path.is_ok() {
+            Err(TestFailure::invalid(case))
+        } else {
+            Ok(())
+        }
+    } else {
+        if let Some(doc) = case.document.as_ref() {
+            let js_path = js_path.map_err(|err| (err, case))?;
+            let result = js_path.find(doc);
+
+            match (case.result.as_ref(), case.results.as_ref()) {
+                (Some(expected), _) => {
+                    if result == *expected {
+                        Ok(())
+                    } else {
+                        Err(TestFailure::match_one(case, &result))
+                    }
+                }
+                (None, Some(expected)) => {
+                    if expected.iter().any(|exp| result == *exp) {
+                        Ok(())
+                    } else {
+                        Err(TestFailure::match_any(case, &result))
+                    }
+                }
+                _ => Ok(()),
+            }
+        } else {
+            Ok(())
+        }
+    }
+}
+
 
 #[derive(serde::Deserialize)]
 struct FilterCase {
@@ -93,3 +132,4 @@ impl<'a> TestFailure<'a> {
         )
     }
 }
+
